Updated the main script
**progress update**:
When looking at the code, I realized it most likely won't work. Why? Because it will only run in the editor. Not at run time. So instead of the plug-in passing the camera to the water, we have the water retrieve it from the plugin. So that when it is running on run time, it will instead take it from its current viewport.
The water.gd script itself is still work-in-progress. I'm still figuring out how to get the viewprt texture to work in the editor, but it works fine on run time.